SavingTrigger quality of life changes
To use the SavingTrigger, make a new instance of the class as a global variable, and call functions using the reference. If you create a class every time you call a function from SavingTrigger, you wont save the proper information correctly. Only have one instance of it per .cpp file.

There are comments on each of the functions to describe what they do in the SavingTrigger.h and .cpp files.